Refactor (#147)
* README 내용삭제

* README 내용추가

* Refactor: Layout 컴포넌트 주석추가

* Refactor: useGetAuth 예외처리 변경

* Refactor: Layout style-components 명칭변경

* Refactor: messageModal css 명칭 변경

* Refactor: messageModal 주석추가

* Refactor: messageModal style 변경

* Refactor: header css 수정

* Refactor: login api수정

* None

* Refactor: useGetAuth 주석추가

* Refactor: header model 삭제

* Refactor: search 전체삭제 버튼 추가

* Refactor: Grass 예외처리 변경 및 유틸함수 변경

* Refactor: useGetMyInfo 예외처리 변경

* Refactor myProfile, userProfile api 예외처리 변경

* Refactor: signup 유효성 검사 및 profileRef 적용

* Refactor: useGetDiary mapper 프로퍼티명 변경

* Refactor: useGetCheckName 예외처리 문구 변경

* Refactor: AccountProfile api 예외처리 변경

* Refactor: 구독관련 컴포넌트 변경 및 상태관리

* Refactor: useScroll callback 제거 및 ref추가

* Design: Aside 구독리스트 css 변경:

* Design: 임시 테마 색상 추가

* Refactor: useSubscribe 중복 데이터 처리

* Refactor: useScroll 수정

* Refactor: useDiaryList 예외처리 변경

* Refactor: Search scroll 변경 및 예외처리 변경

* Refactor: mydiary 예외처리 및 스크롤 변경

* Refactor: myLikeDiary scroll변경 및 예외처리 변경

* Refactor: Tap 컴포넌트 렌더링 조건 변경

* Refactor: diaryContent 정규표현식 변경

* Refactor: createDiary profileRef 적용

* Refactor: useDeleteDiary 새로고침 적용

* Refactor: myDiary 렌더링조건 변경

* Refactor: usePostComment ref추가

* Refactor: commentList scroll 적용

* Refactor: Complain 예외처리 및 css 변경

* Refactor: useGetNewNotice Ref추가

* Refactor: newsItem 예외처리 변경

* feat: UrlUpdate 생성

* Refactor: 뷰포트 단위 변경

* Refactor: useUpdateUrl 생성

* Refactor: page접근권한 설정

* Design: 디자인 변경

* Design: 초과글씨 제어

* Design: css 수정및 각 api token예외 추가

* Refactor: myProfile profileRef 추가

* Refactor: 주석삭제

* Refactor

* Update README.md

* Update README.md

+++ src/3_widgets/accountProfile/api/useDeleteAccount.js
@@ -10,6 +10,7 @@ export const useDeleteAccount = ()=> {
     const setMessage = useMessage( state => state.setMessage );
     
     const deleteAccount = () => {
+        if ( !cookieGet("token") ) return setMessage("회원탈퇴는 본인만 가능합니다");
         baseFetch("account", {method:"DELETE"}, cookieGet("token"));
     };
 
@@ -19,11 +20,11 @@ export const useDeleteAccount = ()=> {
         switch ( fetchData.status ) {
             case 200:
                 cookieRemove();
-                setMessage("이용해주셔서 감사합니다\n로그인화면으로 이동합니다",loginRoute);
+                setMessage("이용해주셔서 감사합니다\n로그인화면으로 이동합니다",window.location.href = '/login');
                 break;
 
             case 401:
-                setMessage("탈퇴할수 없는 계정입니다.");
+                setMessage("회원탈퇴는 본인만 가능합니다");
                 break;
 
             case 500:
